feat(ui): smart start view improvements
- Add a way to easily switch to the node attached to a specific provisioning entry (either by opening a modal there or a link to control panel)
- Add an info panel to explain some useful info about smart start entries
- Fill node manufacturer/model/protool informations when missing from QR code but existing nodeId
