+++ src/components/nodes-table/AssociationGroups.vue
@@ -135,7 +135,7 @@ export default {
 		async getAssociations(ask = false) {
 			let refresh = false
 			if (ask && this.node.status !== 'Dead') {
-				refresh = await this.$listeners.showConfirm(
+				refresh = await this.app.confirm(
 					'Info',
 					`Do you want to force query associations?${
 						this.node.status === 'Alive'
@@ -218,7 +218,7 @@ export default {
 			const args = [this.node.id]
 
 			if (
-				!(await this.$listeners.showConfirm(
+				!(await this.app.confirm(
 					'Attention',
 					`Are you sure you want to remove all associations from this node? This will also remove lifeline association if it exists.`,
 					'alert',
